requirements.txt
# Core dependencies
neo4j>=5.14.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
psutil>=5.9.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
